pyproject.toml
[project]
name = "ai-code-reviewer"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "click>=8.2.1",
    "openai>=1.93.0",
    "python-dotenv>=1.1.1",
    "pyyaml>=6.0.2",
    "requests>=2.32.4",
]

[dependency-groups]
dev = [
    "bandit>=1.8.5",
    "black>=25.1.0",
    "flake8>=7.3.0",
    "pytest>=8.4.1",
    "radon>=6.0.1",
]
